Update build configuration to reduce friction
On systems with glibc < 2.34, the `-lpthread` option needs to be
included when linking with Google's benchmark and test frameworks.
However, the `pkg-config` command for `benchmark` needs to be explicitly
given the `--static` flag for this to be included. The static inclusion
has also been added to the Meson configuration.

I've also included some rudimentary logic to allow the Makefile to
shrink the scope of its compilation targets when an older compiler is
being used. I changed the benchmarks to run using `benchmark_main` in
the same way that the tests use `gtest_main`. Finally, I turned off
compilation of the package-internal tests for Google Benchmarks when
running in GitHub actions.