Optimise partial sort to use a single pass
The original partial sort function was a combination of a QuickSelect
pass followed by a QuickSort pass. This can lead to unnecessary
comparisons and movement of data. This update changes the internal
method used by `avx512_partial_qsort` so that it can more efficiently
sort or recurse into each partition as required and without needing to
take a second sweep over the array. This commit also tweaks the
conditional checks in `avx512_qselect` to avoid a redundant comparison.