Robustify CXX detection logic
The first line wasn't doing anything previously because CXX is defined
as `g++` by default. This change means that now when CXX is undefined or
empty it can be detected automatically. Additionally, the variable is
marked as exported so that the detected version will be passed on to
meson if it is being invoked via make.